Implement event coalescing with shm trigger
The big issue with 576a4e8 (EventTracker) is that I fundamentally misunderstood
how zmtrigger.pl works. I assumed it's request/response, but it isn't. zmtrigger.pl
takes commands and outputs a stream of events regardless of those commands.

This meant that *usually* zmtrigger.pl would, if you used it to trigger an alarm,
give you line telling you about the event you just created. But not always,
if another monitor happened to be triggered you'd get that event ID instead.
In particular a frequent situation was that re-triggering during an ongoing event
would result in zmtrigger.pl handing out the event ID for a linked monitor,
which caused EventTracker to go back and forth between the event ID on the real
monitor and the event ID on the linked monitor. That also meant EventTracker would
end up editing Events.Notes for events on monitos that zm-aidect instance wasn't even
attached to! Oops.

All of this is fixed by directly triggering zmc using shared memory, then waiting
for zmc to enter Alert/Alarm state and retrieving the last_event_id. That seems
to be fairly fool-proof.

Instead of using a timeout we should really be watching for an *->Idle state transition
and then button up the event notes, but that can be done later.